+++ changelog.md
@@ -7,6 +7,7 @@
 
 ### Changes
 * Mob fake sounds now set the current player block pos so it always plays in the same area even if the player leaves
+* Bows now accept luck enchantment
 
 ## 4.8.1
 * Startup crash fix